+++ lib/fastly/models/historical_field_aggregate_response.rb
@@ -44,7 +44,7 @@ module Fastly
         :'status' => :'String',
         :'meta' => :'HistoricalMeta',
         :'msg' => :'String',
-        :'data' => :'Array<Hash<String, String>>'
+        :'data' => :'Array<HistoricalFieldResultsAttributes>'
       }
     end