+++ lib/fastly/models/historical_field_results_attributes_all_of.rb
@@ -12,13 +12,16 @@ require 'date'
 require 'time'
 
 module Fastly
-  class HistoricalFieldAggregateResponseAllOf
-    attr_accessor :data
+  class HistoricalFieldResultsAttributesAllOf
+    attr_accessor :service_id
+
+    attr_accessor :start_time
 
     # Attribute mapping from ruby-style variable name to JSON key.
     def self.attribute_map
       {
-        :'data' => :'data'
+        :'service_id' => :'service_id',
+        :'start_time' => :'start_time'
       }
     end
 
@@ -30,13 +33,15 @@ module Fastly
     # Attribute type mapping.
     def self.fastly_types
       {
-        :'data' => :'Array<HistoricalFieldResultsAttributes>'
+        :'service_id' => :'String',
+        :'start_time' => :'Integer'
       }
     end
 
     # List of attributes with nullable: true
     def self.fastly_nullable
       Set.new([
+        :'service_id',
       ])
     end
 
@@ -44,21 +49,23 @@ module Fastly
     # @param [Hash] attributes Model attributes in the form of hash
     def initialize(attributes = {})
       if (!attributes.is_a?(Hash))
-        fail ArgumentError, "The input argument (attributes) must be a hash in `Fastly::HistoricalFieldAggregateResponseAllOf` initialize method"
+        fail ArgumentError, "The input argument (attributes) must be a hash in `Fastly::HistoricalFieldResultsAttributesAllOf` initialize method"
       end
 
       # check to see if the attribute exists and convert string to symbol for hash key
       attributes = attributes.each_with_object({}) { |(k, v), h|
         if (!self.class.attribute_map.key?(k.to_sym))
-          fail ArgumentError, "`#{k}` is not a valid attribute in `Fastly::HistoricalFieldAggregateResponseAllOf`. Please check the name to make sure it's valid. List of attributes: " + self.class.attribute_map.keys.inspect
+          fail ArgumentError, "`#{k}` is not a valid attribute in `Fastly::HistoricalFieldResultsAttributesAllOf`. Please check the name to make sure it's valid. List of attributes: " + self.class.attribute_map.keys.inspect
         end
         h[k.to_sym] = v
       }
 
-      if attributes.key?(:'data')
-        if (value = attributes[:'data']).is_a?(Array)
-          self.data = value
-        end
+      if attributes.key?(:'service_id')
+        self.service_id = attributes[:'service_id']
+      end
+
+      if attributes.key?(:'start_time')
+        self.start_time = attributes[:'start_time']
       end
     end
 
@@ -80,7 +87,8 @@ module Fastly
     def ==(o)
       return true if self.equal?(o)
       self.class == o.class &&
-          data == o.data
+          service_id == o.service_id &&
+          start_time == o.start_time
     end
 
     # @see the `==` method
@@ -92,7 +100,7 @@ module Fastly
     # Calculates hash code according to all attributes.
     # @return [Integer] Hash code
     def hash
-      [data].hash
+      [service_id, start_time].hash
     end
 
     # Builds the object from hash