+++ lib/fastly/models/realtime_measurements.rb
@@ -71,6 +71,9 @@ module Fastly
     # The total, actual amount of time used to process your requests, including active CPU time (in milliseconds).
     attr_accessor :compute_request_time_ms
 
+    # The total amount of request processing time you will be billed for, measured in 50 millisecond increments.
+    attr_accessor :compute_request_time_billed_ms
+
     # Number of requests from edge to the shield POP.
     attr_accessor :shield
 
@@ -608,17 +611,17 @@ module Fastly
     # Total published messages sent to end users.
     attr_accessor :fanout_send_publishes
 
-    # The total number of class a operations for the object store.
+    # The total number of class a operations for the KV store.
+    attr_accessor :kv_store_class_a_operations
+
+    # The total number of class b operations for the KV store.
+    attr_accessor :kv_store_class_b_operations
+
+    # Use kv_store_class_a_operations.
     attr_accessor :object_store_class_a_operations
 
-    # The total number of class b operations for the object store.
+    # Use kv_store_class_b_operations.
     attr_accessor :object_store_class_b_operations
-
-    # Use object_store_class_b_operations.
-    attr_accessor :object_store_read_requests
-
-    # Use object_store_class_a_operations.
-    attr_accessor :object_store_write_requests
 
     # Total header bytes received from end users over Fanout connections.
     attr_accessor :fanout_req_header_bytes
@@ -687,6 +690,7 @@ module Fastly
         :'compute_execution_time_ms' => :'compute_execution_time_ms',
         :'compute_ram_used' => :'compute_ram_used',
         :'compute_request_time_ms' => :'compute_request_time_ms',
+        :'compute_request_time_billed_ms' => :'compute_request_time_billed_ms',
         :'shield' => :'shield',
         :'ipv6' => :'ipv6',
         :'imgopto' => :'imgopto',
@@ -866,10 +870,10 @@ module Fastly
         :'websocket_resp_body_bytes' => :'websocket_resp_body_bytes',
         :'fanout_recv_publishes' => :'fanout_recv_publishes',
         :'fanout_send_publishes' => :'fanout_send_publishes',
+        :'kv_store_class_a_operations' => :'kv_store_class_a_operations',
+        :'kv_store_class_b_operations' => :'kv_store_class_b_operations',
         :'object_store_class_a_operations' => :'object_store_class_a_operations',
         :'object_store_class_b_operations' => :'object_store_class_b_operations',
-        :'object_store_read_requests' => :'object_store_read_requests',
-        :'object_store_write_requests' => :'object_store_write_requests',
         :'fanout_req_header_bytes' => :'fanout_req_header_bytes',
         :'fanout_req_body_bytes' => :'fanout_req_body_bytes',
         :'fanout_resp_header_bytes' => :'fanout_resp_header_bytes',
@@ -915,6 +919,7 @@ module Fastly
         :'compute_execution_time_ms' => :'Float',
         :'compute_ram_used' => :'Integer',
         :'compute_request_time_ms' => :'Float',
+        :'compute_request_time_billed_ms' => :'Float',
         :'shield' => :'Integer',
         :'ipv6' => :'Integer',
         :'imgopto' => :'Integer',
@@ -1094,10 +1099,10 @@ module Fastly
         :'websocket_resp_body_bytes' => :'Integer',
         :'fanout_recv_publishes' => :'Integer',
         :'fanout_send_publishes' => :'Integer',
+        :'kv_store_class_a_operations' => :'Integer',
+        :'kv_store_class_b_operations' => :'Integer',
         :'object_store_class_a_operations' => :'Integer',
         :'object_store_class_b_operations' => :'Integer',
-        :'object_store_read_requests' => :'Integer',
-        :'object_store_write_requests' => :'Integer',
         :'fanout_req_header_bytes' => :'Integer',
         :'fanout_req_body_bytes' => :'Integer',
         :'fanout_resp_header_bytes' => :'Integer',
@@ -1211,6 +1216,10 @@ module Fastly
 
       if attributes.key?(:'compute_request_time_ms')
         self.compute_request_time_ms = attributes[:'compute_request_time_ms']
+      end
+
+      if attributes.key?(:'compute_request_time_billed_ms')
+        self.compute_request_time_billed_ms = attributes[:'compute_request_time_billed_ms']
       end
 
       if attributes.key?(:'shield')
@@ -1929,20 +1938,20 @@ module Fastly
         self.fanout_send_publishes = attributes[:'fanout_send_publishes']
       end
 
+      if attributes.key?(:'kv_store_class_a_operations')
+        self.kv_store_class_a_operations = attributes[:'kv_store_class_a_operations']
+      end
+
+      if attributes.key?(:'kv_store_class_b_operations')
+        self.kv_store_class_b_operations = attributes[:'kv_store_class_b_operations']
+      end
+
       if attributes.key?(:'object_store_class_a_operations')
         self.object_store_class_a_operations = attributes[:'object_store_class_a_operations']
       end
 
       if attributes.key?(:'object_store_class_b_operations')
         self.object_store_class_b_operations = attributes[:'object_store_class_b_operations']
-      end
-
-      if attributes.key?(:'object_store_read_requests')
-        self.object_store_read_requests = attributes[:'object_store_read_requests']
-      end
-
-      if attributes.key?(:'object_store_write_requests')
-        self.object_store_write_requests = attributes[:'object_store_write_requests']
       end
 
       if attributes.key?(:'fanout_req_header_bytes')
@@ -2043,6 +2052,7 @@ module Fastly
           compute_execution_time_ms == o.compute_execution_time_ms &&
           compute_ram_used == o.compute_ram_used &&
           compute_request_time_ms == o.compute_request_time_ms &&
+          compute_request_time_billed_ms == o.compute_request_time_billed_ms &&
           shield == o.shield &&
           ipv6 == o.ipv6 &&
           imgopto == o.imgopto &&
@@ -2222,10 +2232,10 @@ module Fastly
           websocket_resp_body_bytes == o.websocket_resp_body_bytes &&
           fanout_recv_publishes == o.fanout_recv_publishes &&
           fanout_send_publishes == o.fanout_send_publishes &&
+          kv_store_class_a_operations == o.kv_store_class_a_operations &&
+          kv_store_class_b_operations == o.kv_store_class_b_operations &&
           object_store_class_a_operations == o.object_store_class_a_operations &&
           object_store_class_b_operations == o.object_store_class_b_operations &&
-          object_store_read_requests == o.object_store_read_requests &&
-          object_store_write_requests == o.object_store_write_requests &&
           fanout_req_header_bytes == o.fanout_req_header_bytes &&
           fanout_req_body_bytes == o.fanout_req_body_bytes &&
           fanout_resp_header_bytes == o.fanout_resp_header_bytes &&
@@ -2252,7 +2262,7 @@ module Fastly
     # Calculates hash code according to all attributes.
     # @return [Integer] Hash code
     def hash
-      [requests, logging, log, resp_header_bytes, header_size, resp_body_bytes, body_size, hits, miss, pass, synth, errors, hits_time, miss_time, miss_histogram, compute_requests, compute_execution_time_ms, compute_ram_used, compute_request_time_ms, shield, ipv6, imgopto, imgopto_shield, imgopto_transforms, otfp, otfp_shield, otfp_manifests, video, pci, http2, http3, restarts, req_header_bytes, req_body_bytes, bereq_header_bytes, bereq_body_bytes, waf_blocked, waf_logged, waf_passed, attack_req_header_bytes, attack_req_body_bytes, attack_resp_synth_bytes, attack_logged_req_header_bytes, attack_logged_req_body_bytes, attack_blocked_req_header_bytes, attack_blocked_req_body_bytes, attack_passed_req_header_bytes, attack_passed_req_body_bytes, shield_resp_header_bytes, shield_resp_body_bytes, otfp_resp_header_bytes, otfp_resp_body_bytes, otfp_shield_resp_header_bytes, otfp_shield_resp_body_bytes, otfp_shield_time, otfp_deliver_time, imgopto_resp_header_bytes, imgopto_resp_body_bytes, imgopto_shield_resp_header_bytes, imgopto_shield_resp_body_bytes, status_1xx, status_2xx, status_3xx, status_4xx, status_5xx, status_200, status_204, status_206, status_301, status_302, status_304, status_400, status_401, status_403, status_404, status_406, status_416, status_429, status_500, status_501, status_502, status_503, status_504, status_505, uncacheable, pass_time, tls, tls_v10, tls_v11, tls_v12, tls_v13, object_size_1k, object_size_10k, object_size_100k, object_size_1m, object_size_10m, object_size_100m, object_size_1g, object_size_other, recv_sub_time, recv_sub_count, hash_sub_time, hash_sub_count, miss_sub_time, miss_sub_count, fetch_sub_time, fetch_sub_count, pass_sub_time, pass_sub_count, pipe_sub_time, pipe_sub_count, deliver_sub_time, deliver_sub_count, error_sub_time, error_sub_count, hit_sub_time, hit_sub_count, prehash_sub_time, prehash_sub_count, predeliver_sub_time, predeliver_sub_count, hit_resp_body_bytes, miss_resp_body_bytes, pass_resp_body_bytes, compute_req_header_bytes, compute_req_body_bytes, compute_resp_header_bytes, compute_resp_body_bytes, imgvideo, imgvideo_frames, imgvideo_resp_header_bytes, imgvideo_resp_body_bytes, imgvideo_shield, imgvideo_shield_frames, imgvideo_shield_resp_header_bytes, imgvideo_shield_resp_body_bytes, log_bytes, edge_requests, edge_resp_header_bytes, edge_resp_body_bytes, origin_revalidations, origin_fetches, origin_fetch_header_bytes, origin_fetch_body_bytes, origin_fetch_resp_header_bytes, origin_fetch_resp_body_bytes, shield_revalidations, shield_fetches, shield_fetch_header_bytes, shield_fetch_body_bytes, shield_fetch_resp_header_bytes, shield_fetch_resp_body_bytes, segblock_origin_fetches, segblock_shield_fetches, compute_resp_status_1xx, compute_resp_status_2xx, compute_resp_status_3xx, compute_resp_status_4xx, compute_resp_status_5xx, edge_hit_requests, edge_miss_requests, compute_bereq_header_bytes, compute_bereq_body_bytes, compute_beresp_header_bytes, compute_beresp_body_bytes, origin_cache_fetches, shield_cache_fetches, compute_bereqs, compute_bereq_errors, compute_resource_limit_exceeded, compute_heap_limit_exceeded, compute_stack_limit_exceeded, compute_globals_limit_exceeded, compute_guest_errors, compute_runtime_errors, edge_hit_resp_body_bytes, edge_hit_resp_header_bytes, edge_miss_resp_body_bytes, edge_miss_resp_header_bytes, origin_cache_fetch_resp_body_bytes, origin_cache_fetch_resp_header_bytes, shield_hit_requests, shield_miss_requests, shield_hit_resp_header_bytes, shield_hit_resp_body_bytes, shield_miss_resp_header_bytes, shield_miss_resp_body_bytes, websocket_req_header_bytes, websocket_req_body_bytes, websocket_resp_header_bytes, websocket_bereq_header_bytes, websocket_bereq_body_bytes, websocket_beresp_header_bytes, websocket_beresp_body_bytes, websocket_conn_time_ms, websocket_resp_body_bytes, fanout_recv_publishes, fanout_send_publishes, object_store_class_a_operations, object_store_class_b_operations, object_store_read_requests, object_store_write_requests, fanout_req_header_bytes, fanout_req_body_bytes, fanout_resp_header_bytes, fanout_resp_body_bytes, fanout_bereq_header_bytes, fanout_bereq_body_bytes, fanout_beresp_header_bytes, fanout_beresp_body_bytes, fanout_conn_time_ms, ddos_action_limit_streams_connections, ddos_action_limit_streams_requests, ddos_action_tarpit_accept, ddos_action_tarpit, ddos_action_close, ddos_action_blackhole].hash
+      [requests, logging, log, resp_header_bytes, header_size, resp_body_bytes, body_size, hits, miss, pass, synth, errors, hits_time, miss_time, miss_histogram, compute_requests, compute_execution_time_ms, compute_ram_used, compute_request_time_ms, compute_request_time_billed_ms, shield, ipv6, imgopto, imgopto_shield, imgopto_transforms, otfp, otfp_shield, otfp_manifests, video, pci, http2, http3, restarts, req_header_bytes, req_body_bytes, bereq_header_bytes, bereq_body_bytes, waf_blocked, waf_logged, waf_passed, attack_req_header_bytes, attack_req_body_bytes, attack_resp_synth_bytes, attack_logged_req_header_bytes, attack_logged_req_body_bytes, attack_blocked_req_header_bytes, attack_blocked_req_body_bytes, attack_passed_req_header_bytes, attack_passed_req_body_bytes, shield_resp_header_bytes, shield_resp_body_bytes, otfp_resp_header_bytes, otfp_resp_body_bytes, otfp_shield_resp_header_bytes, otfp_shield_resp_body_bytes, otfp_shield_time, otfp_deliver_time, imgopto_resp_header_bytes, imgopto_resp_body_bytes, imgopto_shield_resp_header_bytes, imgopto_shield_resp_body_bytes, status_1xx, status_2xx, status_3xx, status_4xx, status_5xx, status_200, status_204, status_206, status_301, status_302, status_304, status_400, status_401, status_403, status_404, status_406, status_416, status_429, status_500, status_501, status_502, status_503, status_504, status_505, uncacheable, pass_time, tls, tls_v10, tls_v11, tls_v12, tls_v13, object_size_1k, object_size_10k, object_size_100k, object_size_1m, object_size_10m, object_size_100m, object_size_1g, object_size_other, recv_sub_time, recv_sub_count, hash_sub_time, hash_sub_count, miss_sub_time, miss_sub_count, fetch_sub_time, fetch_sub_count, pass_sub_time, pass_sub_count, pipe_sub_time, pipe_sub_count, deliver_sub_time, deliver_sub_count, error_sub_time, error_sub_count, hit_sub_time, hit_sub_count, prehash_sub_time, prehash_sub_count, predeliver_sub_time, predeliver_sub_count, hit_resp_body_bytes, miss_resp_body_bytes, pass_resp_body_bytes, compute_req_header_bytes, compute_req_body_bytes, compute_resp_header_bytes, compute_resp_body_bytes, imgvideo, imgvideo_frames, imgvideo_resp_header_bytes, imgvideo_resp_body_bytes, imgvideo_shield, imgvideo_shield_frames, imgvideo_shield_resp_header_bytes, imgvideo_shield_resp_body_bytes, log_bytes, edge_requests, edge_resp_header_bytes, edge_resp_body_bytes, origin_revalidations, origin_fetches, origin_fetch_header_bytes, origin_fetch_body_bytes, origin_fetch_resp_header_bytes, origin_fetch_resp_body_bytes, shield_revalidations, shield_fetches, shield_fetch_header_bytes, shield_fetch_body_bytes, shield_fetch_resp_header_bytes, shield_fetch_resp_body_bytes, segblock_origin_fetches, segblock_shield_fetches, compute_resp_status_1xx, compute_resp_status_2xx, compute_resp_status_3xx, compute_resp_status_4xx, compute_resp_status_5xx, edge_hit_requests, edge_miss_requests, compute_bereq_header_bytes, compute_bereq_body_bytes, compute_beresp_header_bytes, compute_beresp_body_bytes, origin_cache_fetches, shield_cache_fetches, compute_bereqs, compute_bereq_errors, compute_resource_limit_exceeded, compute_heap_limit_exceeded, compute_stack_limit_exceeded, compute_globals_limit_exceeded, compute_guest_errors, compute_runtime_errors, edge_hit_resp_body_bytes, edge_hit_resp_header_bytes, edge_miss_resp_body_bytes, edge_miss_resp_header_bytes, origin_cache_fetch_resp_body_bytes, origin_cache_fetch_resp_header_bytes, shield_hit_requests, shield_miss_requests, shield_hit_resp_header_bytes, shield_hit_resp_body_bytes, shield_miss_resp_header_bytes, shield_miss_resp_body_bytes, websocket_req_header_bytes, websocket_req_body_bytes, websocket_resp_header_bytes, websocket_bereq_header_bytes, websocket_bereq_body_bytes, websocket_beresp_header_bytes, websocket_beresp_body_bytes, websocket_conn_time_ms, websocket_resp_body_bytes, fanout_recv_publishes, fanout_send_publishes, kv_store_class_a_operations, kv_store_class_b_operations, object_store_class_a_operations, object_store_class_b_operations, fanout_req_header_bytes, fanout_req_body_bytes, fanout_resp_header_bytes, fanout_resp_body_bytes, fanout_bereq_header_bytes, fanout_bereq_body_bytes, fanout_beresp_header_bytes, fanout_beresp_body_bytes, fanout_conn_time_ms, ddos_action_limit_streams_connections, ddos_action_limit_streams_requests, ddos_action_tarpit_accept, ddos_action_tarpit, ddos_action_close, ddos_action_blackhole].hash
     end
 
     # Builds the object from hash